+++ test/WebSites/FormatterWebSite/Startup.cs
@@ -4,7 +4,7 @@
 using System.IO;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Mvc.ModelBinding.Validation;
+using Microsoft.AspNetCore.Mvc.ModelBinding;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace FormatterWebSite
@@ -15,8 +15,8 @@ namespace FormatterWebSite
         {
             services.AddMvc(options =>
             {
-                options.ModelMetadataDetailsProviders.Add(new ValidationExcludeFilter(typeof(Developer)));
-                options.ModelMetadataDetailsProviders.Add(new ValidationExcludeFilter(typeof(Supplier)));
+                options.ModelMetadataDetailsProviders.Add(new SuppressChildValidationMetadataProvider(typeof(Developer)));
+                options.ModelMetadataDetailsProviders.Add(new SuppressChildValidationMetadataProvider(typeof(Supplier)));
 
                 options.InputFormatters.Add(new StringInputFormatter());
             })
